# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/scans": {
      "post": {
        "description": "Queues a new axe-core scan run in one of three modes: single URL, URL list, or crawl. Processing runs asynchronously in the background. Poll GET /scans/:id for status and results.",
        "operationId": "createScan",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CreateSingleUrlScanDto"
                  },
                  {
                    "$ref": "#/components/schemas/CreateUrlListScanDto"
                  },
                  {
                    "$ref": "#/components/schemas/CreateCrawlScanDto"
                  }
                ],
                "discriminator": {
                  "propertyName": "mode",
                  "mapping": {
                    "single_url": "#/components/schemas/CreateSingleUrlScanDto",
                    "url_list": "#/components/schemas/CreateUrlListScanDto",
                    "crawl": "#/components/schemas/CreateCrawlScanDto"
                  }
                }
              },
              "examples": {
                "singleUrl": {
                  "summary": "Single URL run",
                  "description": "Analyze one page URL.",
                  "value": {
                    "mode": "single_url",
                    "url": "https://example.com",
                    "scanOptions": {
                      "rootElement": "main",
                      "basicAuth": {
                        "username": "scanner-user",
                        "password": "scanner-password"
                      }
                    }
                  }
                },
                "urlList": {
                  "summary": "URL list run",
                  "description": "Analyze a fixed list of URLs.",
                  "value": {
                    "mode": "url_list",
                    "urls": [
                      "https://example.com",
                      "https://example.com/about"
                    ],
                    "scanOptions": {
                      "ruleIds": [
                        "image-alt",
                        "color-contrast"
                      ]
                    }
                  }
                },
                "crawl": {
                  "summary": "Crawl run",
                  "description": "Discover and analyze pages by crawling from one or more seeds.",
                  "value": {
                    "mode": "crawl",
                    "startUrls": [
                      "https://example.com"
                    ],
                    "scanOptions": {
                      "rootElement": "main"
                    },
                    "crawlOptions": {
                      "maxPages": 250,
                      "maxDepth": 4,
                      "strategy": "same-hostname",
                      "globs": [
                        "https://example.com/docs/**"
                      ],
                      "excludeGlobs": [
                        "**/private/**"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Scan run created and queued",
            "headers": {
              "Location": {
                "description": "URL of the created scan run resource",
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Start a new accessibility scan run",
        "tags": [
          "Scans"
        ]
      },
      "get": {
        "description": "Returns all scan runs ordered by creation date, newest first. Use the optional `target` query parameter to filter runs by one of their input targets.",
        "operationId": "listScans",
        "parameters": [
          {
            "name": "target",
            "required": false,
            "in": "query",
            "description": "Filter runs by exact match in normalized input targets. When omitted, all scan runs are returned.",
            "schema": {
              "format": "uri",
              "example": "https://example.com",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of scans",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScanResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all scan runs",
        "tags": [
          "Scans"
        ]
      }
    },
    "/v1/scans/{id}": {
      "get": {
        "description": "Returns a specific scan with full violation details.",
        "operationId": "getScanById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Scan ID",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "pageUrls",
            "required": false,
            "in": "query",
            "description": "Filter returned violations to those with at least one issue on any of the given page URLs. Repeat the parameter for multiple values: ?pageUrls=https://a.com&pageUrls=https://b.com",
            "explode": true,
            "schema": {
              "example": [
                "https://example.com/about"
              ],
              "type": "array",
              "items": {
                "type": "string",
                "format": "uri"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scan found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get scan by ID",
        "tags": [
          "Scans"
        ]
      },
      "delete": {
        "description": "Permanently deletes a scan run and all of its stored issues.",
        "operationId": "deleteScan",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Scan ID",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Scan deleted"
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a scan run",
        "tags": [
          "Scans"
        ]
      }
    },
...
  },
  "info": {
    "title": "MindfulAPI",
    "description": "Automated web accessibility scanning API powered by axe-core and Playwright. Built for use with the [mindfula11y](https://github.com/crinis/mindfula11y) TYPO3 extension. Errors are returned as RFC 9457 `application/problem+json`.",
    "version": "0.5.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Current environment"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Set the AUTH_TOKEN environment variable to enable authentication. When authentication is disabled (AUTH_DISABLED=true), the token is ignored.",
        "type": "http"
      }
    },
    "schemas": {
      "BasicAuthCredentialsDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "scanner-user",
            "minLength": 1,
            "maxLength": 256,
            "description": "Username for HTTP Basic Authentication."
          },
          "password": {
            "type": "string",
            "example": "scanner-password",
            "minLength": 1,
            "maxLength": 1024,
            "description": "Password for HTTP Basic Authentication. Stored encrypted at rest."
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "ScanOptionsDto": {
        "type": "object",
        "properties": {
          "rootElement": {
            "type": "string",
            "example": "main",
            "description": "CSS selector to restrict the scan to a specific page region. Scans the entire page when omitted.",
            "minLength": 1
          },
          "ruleIds": {
            "example": [
              "color-contrast",
              "image-alt"
            ],
            "description": "Specific axe rule IDs to run. All rules run when omitted. See GET /rules for available IDs.",
            "uniqueItems": true,
            "minItems": 1,
            "maxItems": 200,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "basicAuth": {
            "description": "Optional HTTP Basic Authentication credentials used during page loading. Credentials are encrypted at rest and never returned in API responses.",
            "allOf": [
              {
                "$ref": "#/components/schemas/BasicAuthCredentialsDto"
              }
            ]
          }
        }
      },
      "CreateSingleUrlScanDto": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "single_url"
            ],
            "example": "single_url"
          },
          "url": {
            "type": "string",
            "example": "https://example.com",
            "format": "uri",
            "description": "Single page URL to analyze."
          },
          "scanOptions": {
            "$ref": "#/components/schemas/ScanOptionsDto"
          }
        },
        "required": [
          "mode",
          "url"
        ]
      },
      "CreateUrlListScanDto": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "url_list"
            ],
            "example": "url_list"
          },
          "urls": {
            "type": "array",
            "example": [
              "https://example.com",
              "https://example.com/about"
            ],
            "items": {
              "type": "string",
              "format": "uri"
            },
            "uniqueItems": true,
            "minItems": 2,
            "description": "Explicit list of page URLs to analyze."
          },
          "scanOptions": {
            "$ref": "#/components/schemas/ScanOptionsDto"
          }
        },
        "required": [
          "mode",
          "urls"
        ]
      },
      "CrawlOptionsDto": {
        "type": "object",
        "properties": {
          "maxPages": {
            "type": "integer",
            "example": 250,
            "description": "Maximum number of pages to crawl and analyze.",
            "default": 250,
            "minimum": 1,
            "maximum": 5000
          },
          "maxDepth": {
            "type": "integer",
            "example": 4,
            "description": "Maximum crawl depth from each start URL.",
            "default": 4,
            "minimum": 0,
            "maximum": 20
          },
          "strategy": {
            "type": "string",
            "enum": [
              "all",
              "same-hostname",
              "same-domain",
              "same-origin"
            ],
            "example": "same-hostname",
            "default": "same-hostname",
            "description": "URL discovery strategy. `same-hostname` restricts to the seed host, `same-domain` allows subdomains, `same-origin` also matches protocol, `all` follows any link."
          },
          "globs": {
            "example": [
              "https://example.com/docs/**"
            ],
            "description": "Glob patterns. A discovered URL must match at least one pattern to be crawled.",
            "uniqueItems": true,
            "minItems": 1,
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "excludeGlobs": {
            "example": [
              "**/private/**"
            ],
            "description": "Glob patterns. URLs matching any pattern are skipped during crawl.",
            "uniqueItems": true,
            "minItems": 1,
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateCrawlScanDto": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "crawl"
            ],
            "example": "crawl"
          },
          "startUrls": {
            "type": "array",
            "example": [
              "https://example.com"
            ],
            "items": {
              "type": "string",
              "format": "uri"
            },
            "uniqueItems": true,
            "minItems": 1,
            "description": "One or more seed URLs used as crawl entry points for page discovery."
          },
          "scanOptions": {
            "$ref": "#/components/schemas/ScanOptionsDto"
          },
          "crawlOptions": {
            "$ref": "#/components/schemas/CrawlOptionsDto"
          }
        },
        "required": [
          "mode",
          "startUrls"
        ]
      },
      "ScanOptionsResponseDto": {
        "type": "object",
        "properties": {
          "rootElement": {
            "type": "string",
            "example": "main",
            "nullable": true,
            "description": "CSS selector used to scope scans. Null means entire page scanning."
          },
          "ruleIds": {
            "example": [
              "color-contrast",
              "image-alt"
            ],
            "nullable": true,
            "description": "Subset of axe rule IDs executed for this run.",
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "rootElement",
          "ruleIds"
        ]
      },
      "CrawlOptionsResponseDto": {
        "type": "object",
        "properties": {
          "maxPages": {
            "type": "integer",
            "example": 250,
            "minimum": 1,
            "maximum": 5000,
            "description": "Maximum number of pages the crawler will analyze."
          },
          "maxDepth": {
            "type": "integer",
            "example": 4,
            "minimum": 0,
            "maximum": 20,
            "description": "Maximum crawl depth from each seed URL."
          },
          "strategy": {
            "type": "string",
            "enum": [
              "all",
              "same-hostname",
              "same-domain",
              "same-origin"
            ],
            "example": "same-hostname",
            "description": "URL discovery strategy used during crawling."
          },
          "globs": {
            "example": [
              "https://example.com/docs/**"
            ],
            "uniqueItems": true,
            "description": "Glob patterns used to include discovered URLs during crawling.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "excludeGlobs": {
            "example": [
              "**/private/**"
            ],
            "uniqueItems": true,
            "description": "Glob patterns used to exclude discovered URLs during crawling.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "maxPages",
          "maxDepth",
          "strategy",
          "globs",
          "excludeGlobs"
        ]
      },
      "ScanProgressResponseDto": {
        "type": "object",
        "properties": {
          "pagesDiscovered": {
            "type": "integer",
            "example": 120,
            "minimum": 0,
            "description": "Number of unique pages discovered for this run."
          },
          "pagesScanned": {
            "type": "integer",
            "example": 100,
            "minimum": 0,
            "description": "Number of pages successfully analyzed."
          },
          "pagesFailed": {
            "type": "integer",
            "example": 3,
            "minimum": 0,
            "description": "Number of pages that failed during processing."
          }
        },
        "required": [
          "pagesDiscovered",
          "pagesScanned",
          "pagesFailed"
        ]
      },
      "RuleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "color-contrast",
            "description": "Axe rule identifier."
          },
          "description": {
            "type": "string",
            "example": "Elements must have sufficient color contrast"
          },
          "helpUrl": {
            "type": "string",
            "example": "https://dequeuniversity.com/rules/axe/4.11/color-contrast?application=axeAPI",
            "description": "Link to remediation guidance on Deque University.",
            "format": "uri",
            "nullable": true
          },
          "tags": {
            "example": [
              "wcag2aa",
              "wcag143",
              "cat.color"
            ],
            "description": "WCAG and category tags for this rule.",
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "description",
          "helpUrl"
        ]
      },
      "IssueResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1,
            "minimum": 1
          },
          "pageUrl": {
            "type": "string",
            "example": "https://example.com/about",
            "description": "Page URL where this issue occurrence was found.",
            "nullable": true,
            "format": "uri"
          },
          "selector": {
            "type": "string",
            "example": ".btn-primary",
            "description": "CSS selector identifying the element with the issue.",
            "nullable": true
          },
          "context": {
            "type": "string",
            "example": "<button class=\"btn-primary\">Submit</button>",
            "description": "HTML snippet of the problematic element.",
            "nullable": true
          }
        },
        "required": [
          "id",
          "pageUrl",
          "selector",
          "context"
        ]
      },
      "ViolationResponseDto": {
        "type": "object",
        "properties": {
          "rule": {
            "$ref": "#/components/schemas/RuleResponseDto"
          },
          "impact": {
            "type": "string",
            "enum": [
              "critical",
              "serious",
              "moderate",
              "minor"
            ],
            "example": "serious",
            "description": "Severity of the violation as reported by axe-core."
          },
          "issues": {
            "minItems": 1,
            "description": "Issue occurrences for this rule+impact pair.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IssueResponseDto"
            }
          }
        },
        "required": [
          "rule",
          "impact",
          "issues"
        ]
      },
      "ScanResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1,
            "minimum": 1,
            "description": "Unique scan run identifier."
          },
          "mode": {
            "type": "string",
            "enum": [
              "single_url",
              "url_list",
              "crawl"
            ],
            "example": "crawl"
          },
          "targets": {
            "type": "array",
            "example": [
              "https://example.com"
            ],
            "items": {
              "type": "string",
              "format": "uri"
            },
            "uniqueItems": true,
            "minItems": 1,
            "description": "Input URL targets or crawl seeds for this run."
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "completed",
              "failed"
            ],
            "example": "completed"
          },
          "scanOptions": {
            "$ref": "#/components/schemas/ScanOptionsResponseDto"
          },
          "crawlOptions": {
            "nullable": true,
            "description": "Present only for crawl runs.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CrawlOptionsResponseDto"
              }
            ]
          },
          "progress": {
            "description": "Runtime page-processing counters for this run.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ScanProgressResponseDto"
              }
            ]
          },
          "violations": {
            "description": "Accessibility violations grouped by rule and impact severity.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ViolationResponseDto"
            }
          },
          "totalIssueCount": {
            "type": "integer",
            "example": 3,
            "minimum": 0,
            "description": "Sum of issue occurrences across all violations."
          },
          "createdAt": {
            "type": "string",
            "example": "2025-06-14T10:30:00.000Z",
            "format": "date-time",
            "description": "Timestamp when this run was created."
          },
          "updatedAt": {
            "type": "string",
            "example": "2025-06-14T10:31:00.000Z",
            "format": "date-time",
            "description": "Timestamp when this run was last updated."
          }
        },
        "required": [
          "id",
          "mode",
          "targets",
          "status",
          "scanOptions",
          "crawlOptions",
          "progress",
          "violations",
          "totalIssueCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "ValidationErrorDetail": {
        "type": "object",
        "properties": {
          "pointer": {
            "type": "string",
            "example": "/urls",
            "description": "JSON Pointer to the field that failed validation."
          },
          "message": {
            "type": "string",
            "example": "urls must contain at least 2 elements",
            "description": "Human-readable validation message."
          }
        },
        "required": [
          "pointer",
          "message"
        ]
      },
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "about:blank",
            "description": "A URI reference identifying the problem type. `about:blank` is used for problems with no dedicated documentation."
          },
          "title": {
            "type": "string",
            "example": "Bad Request",
            "description": "Short, human-readable summary of the problem type."
          },
          "status": {
            "type": "integer",
            "example": 400,
            "minimum": 400,
            "maximum": 599,
            "description": "HTTP status code for this occurrence."
          },
          "detail": {
            "type": "string",
            "example": "Scan with ID 42 not found",
            "description": "Human-readable explanation specific to this occurrence."
          },
          "instance": {
            "type": "string",
            "example": "/v1/scans/42",
            "description": "URI reference identifying this specific occurrence."
          },
          "errors": {
            "description": "Field-level validation errors (validation problems only).",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorDetail"
            }
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    },
    {}
  ]
}